Fixes ServerlessSnapshotShutdownIT
Clears the shutdown metadata at the end of each test

Relates: #140753
Relates: #140755
Relates: #140759
Relates: #140760
Relates: #140761
Relates: #140762
Relates: #140764
Relates: #140765
