+++ server/src/internalClusterTest/java/org/elasticsearch/snapshots/SnapshotShutdownIT.java
@@ -52,6 +52,7 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 
 import static org.elasticsearch.snapshots.SnapshotShutdownProgressTracker.SNAPSHOT_PROGRESS_DURING_SHUTDOWN_LOG_INTERVAL_SETTING;
+import static org.elasticsearch.snapshots.SnapshotTestUtils.clearShutdownMetadata;
 import static org.elasticsearch.snapshots.SnapshotTestUtils.flushMasterQueue;
 import static org.elasticsearch.snapshots.SnapshotTestUtils.putShutdownForRemovalMetadata;
 import static org.elasticsearch.snapshots.SnapshotTestUtils.putShutdownMetadata;
