+++ test/framework/src/main/java/org/elasticsearch/snapshots/SnapshotTestUtils.java
@@ -47,6 +47,26 @@ public class SnapshotTestUtils {
             .addListener(listener);
     }
 
+    /**
+     * Updates the cluster state to mark a node for shutdown by adding or updating its shutdown metadata. Must be cleared at the end
+     * of the test by calling {@link #clearShutdownMetadata(ClusterService)}
+     * <p>
+     * This method submits an unbatched state update task to the provided {@link ClusterService}, which
+     * sets the shutdown metadata for the specified node. The metadata is built using the provided
+     * {@link SingleNodeShutdownMetadata.Builder}. Upon completion, the provided {@link ActionListener}
+     * is notified.
+     * </p>
+     * Note: At the beginning of each test, the test cluster is reset. Therefore, node names can be reused between tests if they manually
+     * start nodes. When a test adds shutdown metadata for its manually started node, <i>it must be cleared at the end of the test by
+     * calling {@link #clearShutdownMetadata(ClusterService)}</i>.
+     * Otherwise, the subsequent test can start a node with the same name which is still shutting down and cannot accept
+     * shards which leads to create index timeout.
+     *
+     * @param clusterService the {@link ClusterService} used to submit the state update task
+     * @param shutdownMetadataBuilder the builder for the node's shutdown metadata
+     * @param nodeName the name of the node to mark for shutdown
+     * @param listener the {@link ActionListener} to be notified when the operation completes or fails
+     */
     public static void putShutdownMetadata(
         ClusterService clusterService,
         SingleNodeShutdownMetadata.Builder shutdownMetadataBuilder,
@@ -99,5 +119,24 @@ public class SnapshotTestUtils {
                 listener.onResponse(null);
             }
         });
+    }
+
+    public static void clearShutdownMetadata(ClusterService clusterService) {
+        safeAwait(listener -> clusterService.submitUnbatchedStateUpdateTask("remove restart marker", new ClusterStateUpdateTask() {
+            @Override
+            public ClusterState execute(ClusterState currentState) {
+                return currentState.copyAndUpdateMetadata(mdb -> mdb.putCustom(NodesShutdownMetadata.TYPE, NodesShutdownMetadata.EMPTY));
+            }
+
+            @Override
+            public void onFailure(Exception e) {
+                fail(e);
+            }
+
+            @Override
+            public void clusterStateProcessed(ClusterState initialState, ClusterState newState) {
+                listener.onResponse(null);
+            }
+        }));
     }
 }
